code refractoring with separate k8s-client package bug fix for #56 & #57
